Support node and browser variations
